# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation"
  },
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "List of categories",
        "parameters": [
          {
            "name": "language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "page",
            "in": "query",
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      }
    },
...
    "/api/admin/categories": {
      "get": {
        "tags": [
          "Admin - Categories"
        ],
        "summary": "Categories list",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "page",
            "in": "query",
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      }
    },
...
  }
}